+++ include/utils.h
@@ -134,9 +134,9 @@ bool		is_open_pair(unsigned char c, t_is_open_pair_op operation);
 bool		is_valid_varname(char *str);
 bool		is_valid_varname_char(char c);
 bool		is_valid_varname_start(char c);
-bool		skip_dollar_brace(char *str, size_t *i, bool is_in_dquote);
-bool		skip_double_quote(char *str, size_t *i);
-bool		skip_single_quote(char *str, size_t *i);
+bool		skip_dollar_brace(char *str, int *i, bool is_in_dquote);
+bool		skip_double_quote(char *str, int *i);
+bool		skip_single_quote(char *str, int *i);
 char		*concat_list_to_string(t_list *list, char *delim);
 
 #endif
